arduino: add files for Arduino platform
Hacked together an Arduino implementation based off the mbed code. A
lot of hardcoded references to “Serial” that needs to be changed for
proper sample use/debugging. Limited testing. The sample AT interface
seems to work on an Arduino Mega 2560. There might be a few memory
issues, especially considering the limited RAM available on Arduino.
The ATInterface.ino sample is over 19 KB compiled.

To compile the sample create a new directory called ATInterface
anywhere. Copy the contents of the arduino sample directory. Run the
build script in the arduino src directory. Go to lib/XBeeArduino and
copy all the files in that directory to the sketch directory you
created. Open the sketch and select Arduino 2560 from the Tools ->
Board menu. Compile. If you’re having issues with time functions you’ll
need to put the Arduino Time library in the libraries folder.

>Works for me.